fix two compiler warnings on x64 build
See: https://github.com/JNechaevsky/russian-doom/commit/7a202471d6a9cd350befab9193aca9db53680fff

+++ src/heretic/p_saveg.c
@@ -123,7 +123,7 @@ void SV_WriteLong(unsigned int val)
 
 void SV_WritePtr(void *ptr)
 {
-    long val = (long) ptr;
+    long val = (long)(intptr_t) ptr;
 
     SV_WriteLong(val & 0xffffffff);
 }
